Moved Paging to dex/data and removed depedencies on execution.
