+++ sql/src/main/java/io/crate/execution/dsl/projection/FetchProjection.java
@@ -27,7 +27,7 @@ import com.google.common.collect.ImmutableMap;
 import io.crate.expression.symbol.Symbol;
 import io.crate.collections.Lists2;
 import io.crate.metadata.TableIdent;
-import io.crate.execution.support.Paging;
+import io.crate.data.Paging;
 import io.crate.planner.ExplainLeaf;
 import io.crate.planner.node.fetch.FetchSource;
 import org.elasticsearch.common.io.stream.StreamOutput;
