+++ sql/src/main/java/io/crate/execution/engine/NodeOperationTreeGenerator.java
@@ -28,7 +28,7 @@ import io.crate.execution.dsl.phases.NodeOperation;
 import io.crate.execution.dsl.phases.NodeOperationTree;
 import io.crate.execution.dsl.phases.UpstreamPhase;
 import io.crate.execution.engine.distribution.DistributingConsumerFactory;
-import io.crate.execution.support.Paging;
+import io.crate.data.Paging;
 import io.crate.planner.ExecutionPlan;
 import io.crate.planner.ExecutionPlanVisitor;
 import io.crate.planner.Merge;
